+++ phileas/__init__.py
@@ -1,5 +1,7 @@
 # flake8: noqa: F401
 
+__version__ = "0.2"
+
 from .factory import (
     ExperimentFactory,
     Loader,