+++ Changelog.md
@@ -4,3 +4,10 @@
 - Basic experiment creation, with iteration through parameters using cartesian
   product.
 - Experiment graph extraction
+
+# 0.2
+
+- Improve iterables handling
+- Add the IterationMethod.UNION iteration method, which allows to search
+  through iterables one at a time
+- Breaks the iterables API : YAML custom datatypes are now used, and the numeric ranges ends are now called start and end.